+++ Assets/ZombieHunter/InputSystem/DevInputSystem.cs
@@ -1,38 +1,33 @@
-﻿using System.Threading.Tasks;
-using Leopotam.Ecs;
+﻿using Leopotam.Ecs;
 using UnityEngine;
 using ZombieHunter.MovementSystem.Components;
 using ZombieHunter.MovementSystem.Events;
 using ZombieHunter.WeaponSystem;
 using ZombieHunter.WeaponSystem.Components;
 
-namespace ZombieHunter.PlayerInputSystem
+namespace ZombieHunter.InputSystem
 {
-    class PlayerInputSystem : IEcsRunSystem
+    public class DevInputSystem : IEcsRunSystem
     {
-        private float _moveX;
-        private float _moveZ;
-
-        private bool _isRotateble = true;
-
-        //Inject
-        private readonly EcsFilter<Tags.Player, DirectionData, ModelData> _movableFilter = null;
         private readonly EcsFilter<Tags.Player, JumpData> _jumpFilter = null;
+        private readonly EcsFilter<Tags.Player, DirectionData, ModelData> _movableFilter = null;
         private readonly EcsFilter<Tags.RightPlayerWeapon, WeaponData> _rightWeaponFilter = null;
         private readonly EcsFilter<Tags.LeftPlayerWeapon, WeaponData> _leftWeaponFilter = null;
-        
-        private readonly PlayerInputConfig _inputConfig = null;
+       
         private readonly EcsStartup.DevelopMode _devMode = null;
+        
+        private float _moveX;
+        private float _moveZ;
         
         public void Run()
         {
-            if (_devMode.Value)
+            if (!_devMode.Value)
             {
                 return;
             }
             
             SetDirection();
-
+            
             foreach (var i in _movableFilter)
             {
                 ref var directionComponent = ref _movableFilter.Get2(i);
@@ -40,28 +35,22 @@ namespace ZombieHunter.PlayerInputSystem
 
                 direction.x = _moveX;
                 direction.z = _moveZ;
-
-                ref var model = ref _movableFilter.Get3(i);
-
-                DirectionModifier(model.ModelTransform);
             }
 
-            if (OVRInput.GetDown(OVRInput.RawButton.RIndexTrigger))
+            if (Input.GetKeyDown(KeyCode.Space))
             {
-                ShootRightHand();
-            } 
-            
-            if (OVRInput.GetDown(OVRInput.RawButton.LIndexTrigger))
+                Jump();
+            }
+            if (Input.GetMouseButtonDown(0))
             {
                 ShootLeftHand();
             }
             
-            if (OVRInput.GetDown(OVRInput.RawButton.B))
+            if (Input.GetMouseButtonDown(1))
             {
-                Jump();   
+                ShootRightHand();
             }
         }
-
         private void ShootLeftHand()
         {
             foreach (var i in _leftWeaponFilter)
@@ -90,37 +79,11 @@ namespace ZombieHunter.PlayerInputSystem
                 entity.Get<JumpEvent>();
             }
         }
-        
+
         private void SetDirection()
         {
-            var axis = OVRInput.Get(OVRInput.RawAxis2D.LThumbstick);
-           _moveX = axis.x;
-           _moveZ = axis.y;
-        }
-
-        private void DirectionModifier(Transform playerTransform)
-        {
-            var axis = OVRInput.Get(OVRInput.RawAxis2D.RThumbstick);
-            if (_isRotateble)
-            {
-                if (axis.x > _inputConfig.TopPrimaryThumbstickRotateLim)
-                {
-                    playerTransform.Rotate(0f,_inputConfig.RotateModifier,0f);
-                    ModiferTimer();
-                } 
-                else if (axis.x < _inputConfig.LowPrimaryThumbstickRotateLim)
-                {
-                    playerTransform.Rotate(0f,-_inputConfig.RotateModifier,0f);
-                    ModiferTimer();
-                }
-            }
-        }
-
-        private async void ModiferTimer()
-        {
-            _isRotateble = false;
-            await Task.Delay(_inputConfig.DelayPerRotate);
-            _isRotateble = true;
+            _moveX = Input.GetAxis("Horizontal");
+            _moveZ = Input.GetAxis("Vertical");
         }
     }
 }
