+++ Assets/ZombieHunter/MainEcsStartup.cs
@@ -9,7 +9,7 @@ namespace ZombieHunter
     {
         private readonly MovementSystemsContainer _movementSystemsContainer = new MovementSystemsContainer();
         private readonly WaveSystemContainer _waveSystemContainer = new WaveSystemContainer();
-        //private readonly WeaponSystemContainer _weaponSystemContainer = new WeaponSystemContainer();
+        private readonly WeaponSystemContainer _weaponSystemContainer = new WeaponSystemContainer();
         private readonly TakeDamageSystemsContainer _takeDamageSystemsContainer = new TakeDamageSystemsContainer();
             
         protected override void AddSystems()
@@ -18,7 +18,7 @@ namespace ZombieHunter
                 
             _movementSystemsContainer.AddSystems(_systems);
             _waveSystemContainer.AddSystems(_systems);
-            //_weaponSystemContainer.AddSystems(_systems);
+            _weaponSystemContainer.AddSystems(_systems);
             _takeDamageSystemsContainer.AddSystems(_systems);
         }
 
@@ -28,7 +28,7 @@ namespace ZombieHunter
                 
             _movementSystemsContainer.AddOneFrameObjects(_systems);
             _waveSystemContainer.AddOneFrameObjects(_systems);
-            //_weaponSystemContainer.AddOneFrameObjects(_systems);
+            _weaponSystemContainer.AddOneFrameObjects(_systems);
             _takeDamageSystemsContainer.AddOneFrameObjects(_systems);
         }
 
@@ -38,7 +38,7 @@ namespace ZombieHunter
                 
             _movementSystemsContainer.AddInjectors(_systems);
             _waveSystemContainer.AddInjectors(_systems);
-            //_weaponSystemContainer.AddInjectors(_systems);
+            _weaponSystemContainer.AddInjectors(_systems);
             _takeDamageSystemsContainer.AddInjectors(_systems);
         }
     }
